requirements.txt
tabulate
python-dotenv
termcolor